[wip] Example for multi-node prefill/decode disaggregated inference
This change is a WIP. The end goal is to have 3 layers of components:
1. A frontend for client requests. This layer will load balance to
   decode nodes. (Not implemented in this change.)
2. A layer of decode nodes that also communicate to (and load balance across)
   downstream prefill nodes. (Partially implemented in this diff,
without proper LB.)
3. A layer of prefill nodes.

Major TODOs:
- Proper load balancing across decode and prefill nodes. For starters,
  we can start with least-loaded selection.
- Make sure this example (in particular, anything related to load
  balancing) works with DP.

Can be run locally for now as follows:
``
$ kill $(lsof -t -i:8000) ; \
    DYN_LOG=debug VLLM_LOGGING_LEVEL=DEBUG DYN_DISABLE_AUTO_GPU_ALLOCATION=1 CUDA_VISIBLE_DEVICES=2 \
   dynamo serve graphs.disagg_decode:Frontend -f configs/disagg_decode.yaml

$ kill $(lsof -t -i:8001) ; \
    DYN_LOG=debug VLLM_LOGGING_LEVEL=DEBUG DYN_DISABLE_AUTO_GPU_ALLOCATION=1 CUDA_VISIBLE_DEVICES=3 \
    dynamo serve graphs.disagg_prefill:Frontend -f configs/disagg_prefill.yaml

$ curl localhost:8000/v1/completions \
  -H "Content-Type: application/json" \
  -d '{
    "model": "deepseek-ai/DeepSeek-R1-Distill-Llama-8B",
    "prompt": "Who won the 2004 World Series? I vaguely remember the
Boston Red Sox were in it, but I cannot remember the other team.
Also, who was MVP, and which player hit the most home runs in the
series?",
    "stream": false,
    "max_tokens": 300
  }'
```

In the current direction, eventually we'll run a single `dynamo serve` command on each
node (each node will play either a prefill or decode role, but not both). Each node will
have vLLM DP enabled for attention layers, as well as host a subset of
EP-deployed experts.